+++ markmeld/__init__.py
@@ -1,8 +1,8 @@
 from .melder import MarkdownMelder
 from .cli import main
-from .utilities import load_config_file
+from .utilities import load_config_file, load_config_wrapper
 
-__all__ = ["MarkdownMelder", "load_config_file"]
+__all__ = ["MarkdownMelder", "load_config_file", "load_config_wrapper"]
 
 if __name__ == "__main__":
     try: